+++ src/pages/MainPage/MainPage.style.ts
@@ -1,0 +1,23 @@
+import styled from '@emotion/styled';
+
+const ContentContainerStyled = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
+
+const BannerWrapperStyled = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 5% 0;
+`;
+
+const EventCardWrapperStyled = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  margin-bottom: 5%;
+`;
+
+export { ContentContainerStyled, BannerWrapperStyled, EventCardWrapperStyled };
