+++ src/pages/MainPage/MainPage.tsx
@@ -6,27 +6,12 @@ import Tab from '@/components/common/Tab/Tab';
 import { TAB_CONSTANTS } from '@/constants/tab';
 import { TabContextProvider } from '@/context/TabContext';
 import { truncateText } from '@/utils/truncateText';
-import styled from '@emotion/styled';
 
-const ContentContainerStyled = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
-const BannerWrapperStyled = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 5% 0;
-`;
-
-const EventCardWrapperStyled = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  margin-bottom: 5%;
-`;
+import {
+  BannerWrapperStyled,
+  ContentContainerStyled,
+  EventCardWrapperStyled,
+} from './MainPage.style';
 
 const MainPage = () => {
   return (
