+++ src/routes/index.tsx
@@ -1,9 +1,8 @@
 import App from '@/App';
-
 import FormLayout from '@/pages/FormLayout/FormLayout';
 import Layout from '@/pages/Layout/Layout';
 import LoginPage from '@/pages/LoginPage/LoginPage';
-import MainPage from '@/pages/MainPage';
+import MainPage from '@/pages/MainPage/MainPage';
 import NotFoundPage from '@/pages/NotFoundPage';
 import OauthRedirectPage from '@/pages/OauthRedirectPage';
 import ProfilePage from '@/pages/ProfilePage';
